refactor: extract helper methods to util
